scripts/requirements.txt
supabase==2.3.4
python-dotenv==1.0.0
openai==1.54.3
pypdf==4.0.1 